requirements.txt
chromadb>=0.4.0
openai>=1.0.0
gitpython>=3.1.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
pathspec>=0.11.0
requests>=2.28.0
pymongo>=4.0.0